Separate oStd - the stuff that will be in compilers real soon - from a new lib called oBase, which is everything else that was in oStd. Create a new namespace "ouro" - short for ouroboros - and unify that with oCore's objects.
In this process there were a few string-related bugs, so this fixed a off-by-1 in mbsltowcs and wcslmbs. Also with such large changes this exposed that spawn's buffer wasn't large enough to keep taking input while processing it, so bump that up.

Why all this? There is generic code to be moved into oBase that relies on more advanced math - oHLSL math to be precise, but oHLSL's C++ implementation depended on oStd for what should be standard C++. This move decouples things such that oHLSL can rely on oStd and oBase can rely on oStd, and then oBase can rely on oHLSL without oHLSL reversely relying on oBase.